Day one creating the stats and inventory system
Once we've reached a stage of wanting to implement gameplay variants for testing we're finding ourselves in need to implement coding strategies, or variants? or architecture? I dont know what its called but the scpecific one is called Strategy Pattern, I'm told its for avoiding messy code. it will allow us to program important variants in how our systems work. Scpecifically what triggered the need for this, is because we don't know yet if we want to go with a Equipment system that restricts (hard codes) specific sets for specific clsses, or leaving it open to the player, with rewards and flexibility for interesing builds. So we'll test later for complexity and the return each gives.

At this point there's no strategy pattern, we will attempt to create it from this point forward, this snapshot functionning as a save state if we fuck it all up.